src/main/test/java/com/aoc/y2015/ProblemDay16Test.java
package com.aoc.y2015;

import com.aoc.ProblemDayTest;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.fest.assertions.Assertions.assertThat;

public class ProblemDay16Test implements ProblemDayTest {

    private ProblemDay16 problem;

    @BeforeEach
    public void before() {
        problem = new ProblemDay16();
    }

    @Test
    @Override
    public void testSolveA() {
        assertThat(problem.solveA()).isEqualTo(373);
    }

    @Test
    @Override
    public void testSolveB() {
        assertThat(problem.solveB()).isEqualTo(260);
    }

}
